# [exporter/file] Improving write performance speed