--- a/exporter/fileexporter/buffered_writer.go
+++ b/exporter/fileexporter/buffered_writer.go
@@ -1,3 +1,17 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package fileexporter
 
 import (
@@ -8,7 +22,7 @@ import (
 )
 
 // bufferedWriteCloser is intended to use more memory
-// in order to optimise writing to disk to help improve performance.
+// in order to optimize writing to disk to help improve performance.
 type bufferedWriteCloser struct {
 	wraped   io.Closer
 	buffered *bufio.Writer