--- a/exporter/fileexporter/buffered_writer_test.go
+++ b/exporter/fileexporter/buffered_writer_test.go
@@ -1,9 +1,22 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package fileexporter
 
 import (
 	"bytes"
 	"io"
-	"io/ioutil"
 	"os"
 	"path"
 	"testing"
@@ -43,10 +56,10 @@ var (
 
 func BenchmarkWriter(b *testing.B) {
 	tempfile := func(tb testing.TB) io.WriteCloser {
-		f, err := ioutil.TempFile(tb.TempDir(), tb.Name())
+		f, err := os.CreateTemp(tb.TempDir(), tb.Name())
 		require.NoError(tb, err, "Must not error when creating benchmark temp file")
 		tb.Cleanup(func() {
-			assert.NoError(tb, os.RemoveAll(path.Dir(f.Name())), "Must clean up files after being writen")
+			assert.NoError(tb, os.RemoveAll(path.Dir(f.Name())), "Must clean up files after being written")
 		})
 		return f
 	}