**Description:** 
Using the library `bufio` to help improve overall performance with the exporter

**Link to tracking Issue:**
N/A

**Testing:** 

I'ver included benchmarks to help verify improvements, on my local these are the expected changes:

```
BenchmarkWriter/buffered-discard-10         	67764800	        17.55 ns/op	      24 B/op	       1 allocs/op
BenchmarkWriter/raw-file-10                 	25841067	        46.40 ns/op	      72 B/op	       2 allocs/op
BenchmarkWriter/buffered-file-10            	74804418	        16.13 ns/op	      24 B/op	       1 allocs/op
BenchmarkWriter/discard-10                  	75218797	        15.92 ns/op	      24 B/op	       1 allocs/op
```

**Documentation:**

N/A, completely transparent to the user.