@MovieStoreGuy 2023-01-16
The latest benchmark results I have from the update are:

```
BenchmarkWriter/discard_10485760_bytes-10         	514807416	         2.222 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/buffered-discard_10485760_bytes-10         	179711494	         6.596 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/raw-file_10485760_bytes-10                 	39634153	        28.55 ns/op	      48 B/op	       1 allocs/op
BenchmarkWriter/buffered-file_10485760_bytes-10            	443848808	         2.717 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/discard_104857600_bytes-10                 	539187712	         2.244 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/buffered-discard_104857600_bytes-10        	182226061	         6.573 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/raw-file_104857600_bytes-10                	37745956	        27.93 ns/op	      48 B/op	       1 allocs/op
BenchmarkWriter/buffered-file_104857600_bytes-10           	448611487	         2.686 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/discard_1073741824_bytes-10                	541026325	         2.208 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/buffered-discard_1073741824_bytes-10       	181645206	         6.629 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/raw-file_1073741824_bytes-10               	39887326	        29.18 ns/op	      48 B/op	       1 allocs/op
BenchmarkWriter/buffered-file_1073741824_bytes-10          	443683308	         2.693 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/discard_10737418240_bytes-10               	530066356	         2.216 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/buffered-discard_10737418240_bytes-10      	182331368	         6.581 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/raw-file_10737418240_bytes-10              	       1	3159734250 ns/op	       0 B/op	       0 allocs/op
BenchmarkWriter/buffered-file_10737418240_bytes-10         	       1	3254270750 ns/op	       0 B/op	       0 allocs/op
```

Which show an increased performance for small writes (bellow the 10MB) but when it is great than the internal buffer then it will end up as the same performance as the raw file.